fix: :zap: Agent template working for researcher
