+++ src/agents/__init__.py
@@ -1,3 +1,4 @@
+from .agents import Agents
 from .agent import Agent
 
 from .planner import Planner
